update site tags
Removes several superfluous tags on the site. Posts should be
categorized by similar tags to group into categories; tags should
not be treated as SEO.

I may revise this at a later date based on discussions with
others. Personally, I find it helpful to C-f the tags page of
a site for interesting content, but with so few blog posts,
people will select on title. Better to have, e.g., a tag of
"cryptography" with all crypto-related posts.